pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lucent"
version = "0.2.0"
description = "MCP server providing memory functionality for LLMs"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Lucent Source Available License 1.0"}
authors = [
    { name = "kahinton" }
]
dependencies = [
    "mcp[cli]>=1.26.0,<2.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.2.2",
    "fastapi>=0.135.1,<1.0",
    "uvicorn>=0.41.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.22,<1.0",
    "bcrypt>=4.0.0",
    "PyNaCl>=1.5.0",
    "docker>=7.0.0",
    "httpx>=0.27.0",
    "cryptography>=46.0.6,<47.0",
    "requests>=2.33.0,<3.0",
    "pyasn1>=0.6.3,<1.0",
    "pygments>=2.20.0,<3.0",
    "PyJWT>=2.12.1,<3.0",
    "langchain>=1.0.0",
    "langchain-core>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-google-genai>=4.0.0",
    "langchain-ollama>=1.0.0",
    "jsonschema>=4.23.0",
]

[project.urls]
Homepage = "https://github.com/kahinton/lucent"
Repository = "https://github.com/kahinton/lucent"
Issues = "https://github.com/kahinton/lucent/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.14.0",
]
ollama = [
    "langchain-ollama>=1.0.0",
]
otel = [
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
    "opentelemetry-instrumentation-fastapi>=0.41b0",
    "opentelemetry-instrumentation-asyncpg>=0.41b0",
    "opentelemetry-instrumentation-logging>=0.41b0",
]

[project.scripts]
lucent = "lucent.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/lucent"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
norecursedirs = ["tests/manual"]
